LCAM-399-calculate-contribution-changes
