+++ crown-court-contribution/src/test/java/uk/gov/justice/laa/crime/contribution/service/CalculateContributionServiceTest.java
@@ -7,8 +7,11 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.justice.laa.crime.contribution.data.builder.TestModelDataBuilder;
 import uk.gov.justice.laa.crime.contribution.dto.ContributionDTO;
+import uk.gov.justice.laa.crime.contribution.model.Contribution;
+import uk.gov.justice.laa.crime.contribution.model.CreateContributionRequest;
 import uk.gov.justice.laa.crime.contribution.staticdata.enums.CaseType;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
 
@@ -24,6 +27,9 @@ class CalculateContributionServiceTest {
     @InjectMocks
     private CalculateContributionService calculateContributionService;
 
+    @Mock
+    private CompareContributionService compareContributionService;
+
     @Test
     void givenAInvalidCaseType_whenCalculateContributionIsInvoked_thenShouldNotCalledCalculateContribution() {
         when(maatCourtDataService.getRepOrderByRepId(anyInt(), anyString())).thenReturn(TestModelDataBuilder.getRepOrderDTO());
@@ -38,6 +44,22 @@ class CalculateContributionServiceTest {
         calculateContributionService.calculateContribution(ContributionDTO.builder().repId(120).caseType(CaseType.APPEAL_CC).build(),
                 TestModelDataBuilder.LAA_TRANSACTION_ID);
         verify(appealContributionService, times(1)).calculateContribution(any(), anyString());
+    }
+
+
+    @Test
+    void givenValidContributionAndCompareResultIsLessThanTwo_whenCreateContribsIsInvoked_thenContributionIsReturn() {
+        when(compareContributionService.compareContribution(any())).thenReturn(1);
+        when(maatCourtDataService.createContribution(any(CreateContributionRequest.class), any())).thenReturn(TestModelDataBuilder.getContribution());
+        Contribution result = calculateContributionService.createContribs(new CreateContributionRequest(), TestModelDataBuilder.LAA_TRANSACTION_ID);
+        assertThat(result).isNotNull();
+    }
+
+    @Test
+    void givenValidContributionAndCompareResultIsGreaterThanTwo_whenCreateContribsIsInvoked_thenNullIsReturn() {
+        when(compareContributionService.compareContribution(any())).thenReturn(3);
+        Contribution result = calculateContributionService.createContribs(new CreateContributionRequest(), TestModelDataBuilder.LAA_TRANSACTION_ID);
+        assertThat(result).isNull();
     }
 
 }
